src/pyproject.toml
[project]
name = "application-performance-metrics"
version = "0.1.0"
description = "Application Performance Metrics"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.2",
    "opentelemetry-distro>=0.50b0",
    "opentelemetry-exporter-otlp>=1.29.0",
    "redis>=5.0.0",
    "pymongo>=4.6.0",
]
